# packages/backend/app/openapi.yaml
openapi: 3.0.3
info:
  title: FinMind API
  version: 0.1.0
  description: API for budgeting, bills, reminders and insights
servers:
  - url: http://localhost:8000
...
paths:
...
  /privacy/export:
    get:
      summary: Download a GDPR personal data export package (ZIP)
      tags: [Privacy]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: ZIP archive with JSON and CSV exports of owned personal data
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /privacy/delete/request:
    post:
      summary: Start irreversible account deletion (password required)
      tags: [Privacy]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, format: password }
            example:
              password: secret123
      responses:
        '200':
          description: Deletion confirmation token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  confirmation_token: { type: string }
                  expires_in_seconds: { type: integer }
                  confirm_phrase: { type: string }
        '400':
          description: Missing password
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Incorrect password or unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /privacy/delete/confirm:
    post:
      summary: Confirm irreversible account deletion
      tags: [Privacy]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmation_token, confirm]
              properties:
                confirmation_token: { type: string }
                confirm: { type: string, enum: [DELETE_MY_DATA] }
            example:
              confirmation_token: abc123
              confirm: DELETE_MY_DATA
      responses:
        '200':
          description: Account and owned personal data permanently deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  deleted_records:
                    type: object
                    additionalProperties: { type: integer }
                  anonymized_audit_logs: { type: integer }
                  deleted_cache_keys: { type: integer }
                  revoked_refresh_sessions: { type: integer }
        '400':
          description: Invalid confirmation or expired token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /privacy/audit:
    get:
      summary: List GDPR privacy audit events for the current user
      tags: [Privacy]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Audit trail entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    action: { type: string }
                    ip_address: { type: string, nullable: true }
                    user_agent: { type: string, nullable: true }
                    details: { type: object, nullable: true }
                    created_at: { type: string, format: date-time }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      example: { error: "invalid credentials" }
...